Move blender icon to top of browse page
